+++ src/instructions/factory_reset.rs
@@ -1,5 +1,5 @@
 use super::{instruction_id, packet_id};
-use crate::Bus;
+use crate::{Bus, Response};
 
 #[repr(u8)]
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
@@ -24,8 +24,8 @@ where
 	/// This will reset all registers to the factory default, including the EEPROM registers.
 	/// The only exceptions are the ID and baud rate settings, which may be kept depending on the `kind` argument.
 	///
-	/// The `motor_id` parameter may be set to [`packet_id::BROADCAST`],
-	/// although the [`Self::broadcast_factory_reset`] is generally easier to use.
+	/// The `motor_id` parameter must not be set to [`packet_id::BROADCAST`],
+	/// Instead use [`Self::broadcast_factory_reset`].
 	///
 	/// Starting with version 42 of the firmware for the MX-series and X-series,
 	/// motors ignore a broadcast reset command with `FactoryResetKind::ResetAll`.
@@ -33,14 +33,10 @@ where
 	/// which would cause multiple motors on the bus to have the same ID.
 	/// At that point, communication with those motors is not possible anymore.
 	/// The only way to restore communication is to physically disconnect all but one motor at a time and re-assign unique IDs.
-	pub fn factory_reset(&mut self, motor_id: u8, kind: FactoryResetKind) -> Result<(), crate::TransferError> {
-		if motor_id == packet_id::BROADCAST {
-			self.broadcast_factory_reset(kind)?;
-		} else {
-			let response = self.transfer_single(motor_id, instruction_id::FACTORY_RESET, 1, |buffer| buffer[0] = kind as u8)?;
-			crate::InvalidParameterCount::check(response.parameters().len(), 0).map_err(crate::ReadError::from)?;
-		}
-		Ok(())
+	/// Or use the ID Inspection Tool in the Dynamixel Wizard 2.0
+	pub fn factory_reset(&mut self, motor_id: u8, kind: FactoryResetKind) -> Result<Response<()>, crate::TransferError> {
+		let response = self.transfer_single(motor_id, instruction_id::FACTORY_RESET, 1, |buffer| buffer[0] = kind as u8)?;
+		Ok(response.try_into()?)
 	}
 
 	/// Reset the settings of all connected motors to the factory defaults.