+++ src/instructions/reboot.rs
@@ -1,5 +1,5 @@
 use super::{instruction_id, packet_id};
-use crate::{Bus, TransferError, WriteError};
+use crate::{Bus, Response, TransferError, WriteError};
 
 impl<ReadBuffer, WriteBuffer> Bus<ReadBuffer, WriteBuffer>
 where
@@ -12,16 +12,11 @@ where
 	/// When a motor reboots, all volatile (non-EEPROM) registers are reset to their initial value.
 	/// This also has the effect of disabling motor torque and resetting the multi-revolution information.
 	///
-	/// The `motor_id` parameter may be set to [`packet_id::BROADCAST`],
-	/// although the [`Self::broadcast_reboot`] is generally easier to use.
-	pub fn reboot(&mut self, motor_id: u8) -> Result<(), TransferError> {
-		if motor_id == packet_id::BROADCAST {
-			self.broadcast_action()?;
-		} else {
-			let response = self.transfer_single(motor_id, instruction_id::REBOOT, 0, |_| ())?;
-			crate::InvalidParameterCount::check(response.parameters().len(), 0).map_err(crate::ReadError::from)?;
-		}
-		Ok(())
+	/// The `motor_id` parameter must not be set to [`packet_id::BROADCAST`],
+	/// Instead use [`Self::broadcast_reboot`].
+	pub fn reboot(&mut self, motor_id: u8) -> Result<Response<()>, TransferError> {
+		let response = self.transfer_single(motor_id, instruction_id::REBOOT, 0, |_| ())?;
+		Ok(response.try_into()?)
 	}
 
 	/// Broadcast an reboot command to all connected motors to trigger a previously registered instruction.