+++ src/instructions/reg_write.rs
@@ -1,5 +1,5 @@
 use super::instruction_id;
-use crate::{Bus, TransferError};
+use crate::{Bus, Response, TransferError};
 
 use crate::endian::{write_u16_le, write_u32_le};
 
@@ -14,13 +14,12 @@ where
 	///
 	/// You can have all connected motors execute their registered write using [`Self::broadcast_action`],
 	/// or a single motor using [`Self::action`].
-	pub fn reg_write(&mut self, motor_id: u8, address: u16, data: &[u8]) -> Result<(), TransferError> {
+	pub fn reg_write(&mut self, motor_id: u8, address: u16, data: &[u8]) -> Result<Response<()>, TransferError> {
 		let response = self.transfer_single(motor_id, instruction_id::REG_WRITE, 2 + data.len(), |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			buffer[2..].copy_from_slice(data)
 		})?;
-		crate::error::InvalidParameterCount::check(response.parameters().len(), 0).map_err(crate::ReadError::from)?;
-		Ok(())
+		Ok(response.try_into()?)
 	}
 
 	/// Register a write command for a 8 bit value to a specific motor.
@@ -29,13 +28,12 @@ where
 	///
 	/// You can have all connected motors execute their registered write using [`Self::broadcast_action`],
 	/// or a single motor using [`Self::action`].
-	pub fn reg_write_u8(&mut self, motor_id: u8, address: u16, value: u8) -> Result<(), TransferError> {
+	pub fn reg_write_u8(&mut self, motor_id: u8, address: u16, value: u8) -> Result<Response<()>, TransferError> {
 		let response = self.transfer_single(motor_id, instruction_id::REG_WRITE, 2 + 1, |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			buffer[2] = value;
 		})?;
-		crate::error::InvalidParameterCount::check(response.parameters().len(), 0).map_err(crate::ReadError::from)?;
-		Ok(())
+		Ok(response.try_into()?)
 	}
 
 	/// Register a write command for a 16 bit value to a specific motor.
@@ -44,13 +42,12 @@ where
 	///
 	/// You can have all connected motors execute their registered write using [`Self::broadcast_action`],
 	/// or a single motor using [`Self::action`].
-	pub fn reg_write_u16(&mut self, motor_id: u8, address: u16, value: u16) -> Result<(), TransferError> {
+	pub fn reg_write_u16(&mut self, motor_id: u8, address: u16, value: u16) -> Result<Response<()>, TransferError> {
 		let response = self.transfer_single(motor_id, instruction_id::REG_WRITE, 2 + 2, |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			write_u16_le(&mut buffer[2..], value);
 		})?;
-		crate::error::InvalidParameterCount::check(response.parameters().len(), 0).map_err(crate::ReadError::from)?;
-		Ok(())
+		Ok(response.try_into()?)
 	}
 
 	/// Register a write command for a 32 bit value to a specific motor.
@@ -59,12 +56,11 @@ where
 	///
 	/// You can have all connected motors execute their registered write using [`Self::broadcast_action`],
 	/// or a single motor using [`Self::action`].
-	pub fn reg_write_u32(&mut self, motor_id: u8, address: u16, value: u32) -> Result<(), TransferError> {
+	pub fn reg_write_u32(&mut self, motor_id: u8, address: u16, value: u32) -> Result<Response<()>, TransferError> {
 		let response = self.transfer_single(motor_id, instruction_id::REG_WRITE, 2 + 4, |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			write_u32_le(&mut buffer[2..], value);
 		})?;
-		crate::error::InvalidParameterCount::check(response.parameters().len(), 0).map_err(crate::ReadError::from)?;
-		Ok(())
+		Ok(response.try_into()?)
 	}
 }